refactor carreers module to mvc
